requirements.txt
nicegui>=2.0.0
pyodbc>=5.1.0
pandas>=2.2.0
python-dotenv>=1.0.1
bcrypt>=4.1.2
passlib>=1.7.4
plotly>=5.22.0
jinja2>=3.1.3
Pillow>=9.0.0
